+++ test/base-queue-handler.test.ts
@@ -2,7 +2,7 @@ import * as should from 'should';
 import * as assert from 'assert';
 import * as sinon from 'sinon';
 import Rabbit from '../ts/rabbit';
-import BaseQueueHandler from '../ts/baseQueueHandler';
+import BaseQueueHandler from '../ts/base-queue-handler';
 const sandbox = sinon.sandbox.create();
 
 describe('Test baseQueueHandler', function() {
