+++ test/delay-queue.test.ts
@@ -1,5 +1,5 @@
 import * as sinon from 'sinon';
-import * as DelayQueue from '../ts/delayQueue';
+import * as DelayQueue from '../ts/delay-queue';
 import Rabbit from '../ts/rabbit';
 import * as Queue from '../ts/queue';
 const sandbox = sinon.sandbox.create();
