+++ ts/index.ts
@@ -1,2 +1,2 @@
 export { default as Rabbit } from './rabbit';
-export { default as BaseQueueHandler } from './baseQueueHandler';
+export { default as BaseQueueHandler } from './base-queue-handler';
